Move all shell commands to their own file and selectively include when frontend is run with Electron
